squashed: implémenter la notation de films
Affichage du bouton selon si admin ou abonne

Déplacer les views abonne dans User

Permettre l'affichage de tous les films à tous les utilisateurs.

Afficher la page liste de film pour l'abonné

Début de l'affichage pour la réservation.

Revert "Début de l'affichage pour la réservation."

This reverts commit 9ce21a416e6cd1ed3cb098b174697230ea767101.

Reapply "Début de l'affichage pour la réservation."

This reverts commit 4ae9f8745810c64a6b7164d92371fbb799978d78.

Affihcage des listes (fonctionnelles sauf films assistés), Détails fonctionnels

+++ CineQuebec.Application/Records/Films/NoteFilmDto.cs
@@ -6,7 +6,7 @@ namespace CineQuebec.Application.Records.Films;
 public record class NoteFilmDto(
     Guid Id,
     FilmDto? IdFilm,
-    ushort Note
+    byte Note
 ) : EntityDto(Id);
 
 internal static class NoteFilmExtensions
